outreach/requirements.txt
streamlit
pandas
google-genai
pypdf
python-docx
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
googlesearch-python
beautifulsoup4
dnspython
